Modificacaiones en el login, integracion en una única clase Conexion
