+++ src/main/java/Main.java
@@ -1,9 +1,9 @@
-import modelo.login.LoginSystem;
+import modelo.login.Conexion;
 
 public class Main {
     public static void main(String[] args) {
         try {
-            LoginSystem.getLoginSystem().logear("admin@gmail.com","123456");
+            Conexion.getInstance().logear("admin@gmail.com","123456");
         } catch (Exception e) {
             e.printStackTrace();
         }
